server/requirements.txt
fastapi==0.111.0
openai==1.30.5
psycopg[binary]==3.1.19
sqlmodel==0.0.18
boto3==1.34.117
botocore==1.34.117
python-dotenv==1.0.1
pandas==2.2.2
uvicorn==0.30.1
simple-ddl-parser==1.5.0
omymodels==0.17.0
datamaker-core==0.0.0